Prioritise indentation from CLI
When the indentation is set from the CLI arguments, respect it.
Otherwise, and by default, use the indentation from the editorconfig
settings, if any. This default behaviour is controlled through a
negative indentation. Whenever no indentation can be found, default to
the constant in the implementation, i.e. 2 (matches the one in lxml).