Reformat on --write only
Reformat the files only when the `--write` option is used. This is made
the default of the hook through its YAML configuration. When `--write`
is not specified, wrongly formatted files will make the hook fail. When
it fails, the hook returns the number of wrongly formatted files+2 -- 2
is necessary to be able to use exit codes that are not reserved (as much
as possible).