Recognise PRE_COMMIT_HOOK_LXML_FORMAT_ env vars
Recognise a number of environment variables led by the
PRE_COMMIT_HOOK_LXML_FORMAT_ prefix. These have **precedence** over the
CLI options. The argument is that the regular case is for the CLI
options to come from the `args` key of the YAML hook configuration.
Since this configuration is for all users of the repository, it is not
appropriate to have the CLI options in the YAML configuration. The
environment variables are for the repository, environment variables are
a way to adapt generic settings to the local host/environment where the
hook is being called.

In those variables, booleans can be represented by human readable
strings, e.g. "yes", "True", "OFF", etc.

Note: 'width' is now renamed to 'indent' to match the CLI option and the
name of the environment variable.